Move DatabaseConnector
