+++ .idea/encodings.xml
@@ -2,8 +2,10 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/analyzer-core/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/analyzer-core/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/analyzer-models/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/analyzer-sql/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/analyzer-sql/src/main/liquibase" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/analyzer-sql/src/main/resources" charset="UTF-8" />
   </component>
 </project>
