+++ analyzer-core/src/main/resources/connection.properties.example
