+++ analyzer-sql/src/main/liquibase/ffb/analyzer/sql/changelog-master.xml
@@ -4,4 +4,6 @@
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
     http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.8.xsd">
+
+    <include file="1.0/changelog.xml" relativeToChangelogFile="true"/>
 </databaseChangeLog>
